Add the source URL for the Dask PyPI packages
To ensure the auto-update bot generates PRs to update the `dask`
feedstock, add the source URL for the PyPI packages. That way it can
detect and update this package when needed. This in spite of the fact
that this is a metapackage that has no build step of its own.