Request bot to check solvable before opening PRs
As `dask` package builds depend on `dask-core` and `distributed` package
builds, if PRs are opened to both feedstocks to update the package
version at the same time, the PR to `dask` will fail until `dask-core`
and `distributed` packages are available. This fixes that issue by
ensuring that `dask`'s dependencies (`dask-core` and `distributed` in
this case) are available before opening a PR. That way the PR will only
be opened after `dask-core` and `distributed` has new packages
published. Should cutdown the time recipe maintainers spend restarting
CI to get the new dependencies.